find by cin

+++ src/main/java/com/example/tnb/dao/IDao.java
@@ -1,5 +1,8 @@
 package com.example.tnb.dao;
 
+import com.example.tnb.entity.Taxe;
+import org.springframework.data.repository.query.Param;
+
 import java.util.List;
 
 public interface IDao<T> {
@@ -9,8 +12,9 @@ public interface IDao<T> {
 
     void delete(T id);
 
-    T findById(int id);
+    T findById(long id);
 
     List<T> findAll();
+
 
 }
